added authentication middleware and created relationship between user and training model, removed form if the user is not logged in, only the user who created that post can edit or delete it.

+++ client/src/components/Card.js
@@ -12,10 +12,18 @@ const Card = ({
   exercises,
   trainerTips,
   editMode,
+  createdBy,
 }) => {
   const dispatch = useDispatch();
+
+  const { _id: userId } = JSON.parse(localStorage.getItem('user'));
+
   const handleChangeDelete = () => {
-    dispatch(deletePost(_id));
+    if (createdBy === userId) {
+      dispatch(deletePost(_id));
+    } else {
+      console.log('Get the fuck out');
+    }
   };
 
   const handleChangeUpdate = () => {
